synthetic-praddle-commit

+++ mlir/lib/Dialect/Bufferization/Transforms/BufferOptimizations.cpp
@@ -122,8 +122,7 @@ static bool defaultIsSmallAlloc(Value alloc, unsigned maximumSizeInBytes,
   if (bitwidth != 0 &&
       *numElements > static_cast<int64_t>(maximumSizeInBytes * 8ULL / bitwidth))
     return false;
-  return *numElements * bitwidth <=
-         static_cast<int64_t>(maximumSizeInBytes * 8ULL);
+  return *numElements * bitwidth <= maximumSizeInBytes * 8;
 }
 
 /// Checks whether the given aliases leave the allocation scope.
